# apis/open-api/drama/drama-app-api.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SDKWork Drama App API",
    "version": "0.1.0",
    "description": "App API contract for SDKWork Drama short-drama episode authoring and media assets. Protected operations use the SDKWork dual-token protocol (Authorization: Bearer <auth_token> plus Access-Token header). All HTTP 2xx JSON responses use the SdkWorkApiResponse envelope ({ code, data, traceId }); all HTTP 4xx/5xx error responses use RFC 9457 application/problem+json with numeric code and traceId. Snowflake int64 entity ids serialize as JSON strings. Authority: ../sdkwork-specs/API_SPEC.md sections 4.5 and 14-16. Contract changes require human review before SDK regeneration.",
    "x-sdkwork-wire-protocol": "sdkwork-v3"
  },
  "servers": [
    {
      "url": "{applicationPublicHttpUrl}",
      "description": "Application public ingress",
      "variables": {
        "applicationPublicHttpUrl": {
          "default": "http://127.0.0.1:8090",
          "description": "Public HTTP URL of the drama application ingress (etc/ topology profiles own real values)."
        }
      }
    }
  ],
...
  "paths": {
...
    "/app/v3/api/episodes/{episodeId}/publish": {
      "post": {
        "tags": [
          "episode"
        ],
        "operationId": "episodes.publish",
        "summary": "Publish a draft episode.",
        "description": "Command operation: only draft episodes can transition to published; other states yield 409 Conflict.",
        "x-sdkwork-api-surface": "app-api",
        "security": [
          {
            "AuthToken": [],
            "AccessToken": []
          }
        ],
        "parameters": [
          {
            "name": "episodeId",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/SnowflakeId"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Episode published.",
            "headers": {
              "X-SdkWork-Trace-Id": {
                "$ref": "#/components/headers/TraceId"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EpisodeResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required or invalid dual token.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found within the caller tenant.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "409": {
            "description": "State conflict (invalid status transition).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "AccessToken": {
        "type": "apiKey",
        "in": "header",
        "name": "Access-Token",
        "description": "SDKWork access token header (IAM_SPEC.md)."
      },
      "AuthToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "SDKWork auth token: Authorization: Bearer <auth_token> (IAM_SPEC.md)."
      }
    },
    "headers": {
      "TraceId": {
        "description": "Server-owned trace id echoed for correlation.",
        "schema": {
          "type": "string"
        }
      }
    },
...
    "schemas": {
      "SnowflakeId": {
        "type": "string",
        "description": "Positive snowflake entity id rendered as a decimal string (SUBJECT_ID_SPEC.md).",
        "pattern": "^[0-9]+$"
      },
...
      "EpisodeStatus": {
        "type": "string",
        "enum": [
          "draft",
          "published",
          "retired"
        ]
      },
...
      "Episode": {
        "type": "object",
        "required": [
          "id",
          "tenantId",
          "title",
          "status",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "$ref": "#/components/schemas/SnowflakeId"
          },
          "tenantId": {
            "$ref": "#/components/schemas/SnowflakeId"
          },
          "title": {
            "type": "string"
          },
          "synopsis": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "$ref": "#/components/schemas/EpisodeStatus"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "EpisodeResponse": {
        "type": "object",
        "required": [
          "code",
          "data",
          "traceId"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "const": 0
          },
          "data": {
            "$ref": "#/components/schemas/Episode"
          },
          "traceId": {
            "type": "string"
          }
        }
      },
...
      "ProblemDetail": {
        "type": "object",
        "description": "RFC 9457 problem+json with SDKWork platform extensions.",
        "required": [
          "type",
          "title",
          "status",
          "code",
          "traceId",
          "instance"
        ],
        "properties": {
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string"
          },
          "code": {
            "type": "integer",
            "description": "Numeric SdkWorkResultCode."
          },
          "traceId": {
            "type": "string"
          },
          "operationId": {
            "type": "string",
            "nullable": true
          },
          "i18nKey": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
    }
  }
}